Javadoc format, VisibleForTest annotation, move telemetry test too
